Fixed two bugs: there was a floating point precision problem on very large references that could be exhibited only on short (e.g. Illumina) reads and default semiglobal mode of alignment. The second problem was a bug in the anchored alignment, where in some rare cases two anchors might overlap. This could have caused some problematic CIGAR strings to appear (i.e. the CIGAR length would be shorter than the read length).

+++ src/containers/path_graph_entry.h
@@ -68,11 +68,11 @@ typedef struct InfoMapping {
 
 typedef struct InfoL1 {
   int64_t l1_l = 0;
-  float l1_k = 1.0f;
+  double l1_k = 1.0f;
   int64_t l1_lmin = 0;
   int64_t l1_lmax = 0;
-  float l1_confidence_abs = 0;
-  float l1_std = 0;
+  double l1_confidence_abs = 0;
+  double l1_std = 0;
   int64_t l1_rough_start = 0;
   int64_t l1_rough_end = 0;
 } InfoL1;
